requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
langchain_community
langchain_openai
faiss-cpu
langchain-groq
